changed the type of the return to size_t instead of int for consistency

+++ src/ft_lstfind.c
@@ -6,17 +6,17 @@
 /*   By: shorwood <shorwood@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/10/15 17:57:57 by shorwood     #+#   ##    ##    #+#       */
-/*   Updated: 2018/12/24 14:38:47 by shorwood    ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/03/28 04:34:40 by shorwood    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_lstfind(t_list **lst, int (*fnc)(void *dat))
+size_t	ft_lstfind(t_list **lst, int (*fnc)(void *dat))
 {
 	t_list	*cur;
-	int		idx;
+	size_t	idx;
 
 	if (!lst || !*lst)
 		return (-1);
